Rearange ROOT and I/O functions
