+++ machine_learning_hep/general.py
@@ -17,12 +17,9 @@ Methods to: load and write data to ROOT files
             filter and manipulate pandas DataFrames
 """
 
-import os
 import uproot
 import yaml
 from pkg_resources import resource_stream
-from ROOT import TFile  # pylint: disable=import-error,no-name-in-module
-from machine_learning_hep.root import fill_ntuple
 
 
 def get_database_ml_parameters():
@@ -39,11 +36,6 @@ def split_df_sigbkg(dataframe_, var_signal_):
     dataframe_sig_ = dataframe_.loc[dataframe_[var_signal_] == 1]
     dataframe_bkg_ = dataframe_.loc[dataframe_[var_signal_] == 0]
     return dataframe_sig_, dataframe_bkg_
-
-
-def checkdir(path):
-    if not os.path.exists(path):
-        os.makedirs(path)
 
 
 def preparestringforuproot(myarray):
@@ -77,11 +69,3 @@ def createstringselection(var_skimming_, minlist_, maxlist_):
     for var, low, high in zip(var_skimming_, minlist_, maxlist_):
         string_selection = string_selection+(("%s_%.1f_%.1f") % (var, low, high))
     return string_selection
-
-
-def write_tree(filename, treename, dataframe):
-    listvar = list(dataframe)
-    values = dataframe.values
-    fout = TFile.Open(filename, "recreate")
-    fout.cd()
-    fill_ntuple(treename, values, listvar)
